Crate internet_reloader

A Rust application to monitor internet connectivity, and if no internet available, attempt to reconnect the network.

This is currently only implemented for Windows OS, as an over-engineered method of resolving internet connectivity issues when using my personal phone’s hotspot for connectivity.

Modules

app

Module for the main application logic.

internet_connectivity

Module for checking internet connectivity.

network_manager

Module for managing network connections.